requirements.txt
pymupdf
python-dotenv
openai